GEMMA_ATTN P2: bf16 block bounds — scoring read halved
Store/read top-k block min/max bounds in the cache dtype (bf16) instead of
fp32. Lossless (min/max of bf16 keys are bf16; maintenance error 0.0) and
halves the bounds read, which is the scoring bottleneck. Decode-attn top-k
speedup 1.28->1.64x (b8/L4096) up to 2.05->2.95x (b32/L8192); >2x across
L8192. Real Gemma4-31B unchanged (3/3). Parallel-warp top-k was tried and
reverted (read-bound, not selection-bound).

+++ vllm/v1/attention/backends/gemma_attn.py
@@ -752,9 +752,11 @@ class GemmaAttentionImpl(AttentionImpl):
             return
         if self.block_bounds is None:
             num_blocks = key_cache.shape[0]
+            # bf16 bounds (== cache dtype): lossless (min/max of bf16 keys) and
+            # halves the scoring read vs fp32.
             self.block_bounds = torch.zeros(
                 num_blocks, 2, self.num_kv_heads, self.actual_head_size,
-                dtype=torch.float32, device=key_cache.device,
+                dtype=key_cache.dtype, device=key_cache.device,
             )
         blocks = (slot // block_size).to(torch.int32)
         offs = (slot % block_size)
